requirements.txt
mysql.connector
datetime
pandas
matplotlib
tabulate